More work on WpfDrawingDocument
- Improved the WpfDrawingDrawing object implementation
- Added support for hit-testing by point and rectangle
- Added a sample, WpfTestSvgControl, for testing this feature
- Modified the codes to make the SetId value same as the SVG ID
- This is the first step in support for interactivities and resolve the issue #106

+++ Samples/WpfTestSvgSample/MainWindow.xaml.cs
@@ -1076,9 +1076,20 @@ namespace WpfTestSvgSample
                     string[] imageFiles = Directory.GetFiles(_drawingDir, "*.png");
                     if (imageFiles != null && imageFiles.Length != 0)
                     {
-                        foreach (var imageFile in imageFiles)
+                        try
                         {
-                            File.Delete(imageFile);
+                            foreach (var imageFile in imageFiles)
+                            {
+                                if (File.Exists(imageFile))
+                                {
+                                    File.Delete(imageFile);
+                                }
+                            }
+                        }
+                        catch (IOException ex)
+                        {
+                            Trace.TraceError(ex.ToString());
+                            // Image this, WPF will typically cache and/or lock loaded images
                         }
                     }
                 }
